requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0
PyMySQL==1.1.0
cryptography==41.0.7
